+++ src/main/java/me/kimihiqq/vouchermanagement/domain/voucherwallet/service/VoucherWalletService.java
@@ -32,7 +32,6 @@ public class VoucherWalletService {
     }
 
 
-
     public void addVoucherToWallet(UUID customerId, UUID voucherId) {
         voucherWalletRepository.addVoucherToWallet(customerId, voucherId);
     }
